# docs/openapi/api.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Swagger Bankdemo
paths:
  /v1/transfer:
    post:
      summary: transfer to another account
      operationId: transferFromTo
      tags:
        - transfer
      requestBody:
        description: account from account to value
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transfer'
      responses:
        '204':
          description: transfered succesfully
        '404':
          description: transfer error
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /v1/accounts/{account}:
    get:
      summary: Info for a specific account
      operationId: exposeAccountById
      tags:
        - accounts
      parameters:
        - name: account
          in: path
          required: true
          description: The account to retrieve
          schema:
            type: long
      responses:
        '200':
          description: Expected response to a valid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        '404':
          description: A specified account was not found.
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    Transfer:
      type: object
      required:
        - from
        - to
        - value
      properties:
        from:
          type: integer
          format: int64
          minimum: 1000000
        to:
          type: integer
          format: int64
          minimum: 1000000
        value:
          type: integer
          format: int64
          minimum: 1
    Account:
      type: object
      required:
        - id
        - account
        - value
      properties:
        id:
          type: integer
          format: int64
        account:
          type: integer
          format: int64
        value:
          type: integer
          format: int64
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string